some fixes after 1st code review
