+++ conveyor/src/main/resources/application.properties
@@ -1,15 +1,5 @@
-baserate=0.12
-
-# logging level
 logging.level.org.springframework=ERROR
 logging.level.com.example.conveyor=DEBUG
-
-# output to a file
 logging.file.name=conveyor.log
-
-# temp folder example
-#logging.file=${java.io.tmpdir}/app.log
-
 logging.pattern.file=%d %p %c{1.} [%t] %m%n
-
 logging.pattern.console=%d{HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n
