--- conveyor/src/test/java/com/example/conveyor/ConveyorApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.conveyor;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ConveyorApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
